+++ app/models/project.rb
@@ -2,6 +2,7 @@ class Project < ActiveRecord::Base
   include GuidReferenced
   acts_as_commentable
   acts_as_authorization_object
+  acts_as_taggable
 
   github_concern :repo => :git_repo_name
 
@@ -44,6 +45,14 @@ class Project < ActiveRecord::Base
 
   def hours
     WorkUnit.for_project(self).sum(:effective_hours)
+  end
+
+  def hours_this_month
+    WorkUnit.for_project(self).scheduled_between(Time.zone.now.beginning_of_month, Time.zone.now.end_of_month).sum(:effective_hours)
+  end
+
+  def hours_this_week
+    WorkUnit.for_project(self).scheduled_between(Time.zone.now.beginning_of_week, Time.zone.now.end_of_week).sum(:effective_hours)
   end
 
   def work_units