+++ app/models/user.rb
@@ -8,6 +8,7 @@ class User < ActiveRecord::Base
   is_gravtastic!
 
   acts_as_authorization_subject :association_name => :roles, :join_table_name => :roles_users
+  acts_as_tagger
 
   # Setup accessible (or protected) attributes for your model
   attr_accessible :email, :password, :password_confirmation, :remember_me,
@@ -66,6 +67,10 @@ class User < ActiveRecord::Base
 
   def to_s
     "#{first_name.capitalize} #{middle_initial.capitalize} #{last_name.capitalize}"
+  end
+
+  def client?
+    @client ||= has_role?(:client)
   end
 
   def developer?