+++ db/schema.rb
@@ -11,34 +11,13 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20120126154403) do
-
-  create_table "audits", :force => true do |t|
-    t.integer  "auditable_id"
-    t.string   "auditable_type"
-    t.integer  "associated_id"
-    t.string   "associated_type"
-    t.integer  "user_id"
-    t.string   "user_type"
-    t.string   "username"
-    t.string   "action"
-    t.text     "audited_changes"
-    t.integer  "version",         :default => 0
-    t.string   "comment"
-    t.string   "remote_address"
-    t.datetime "created_at"
-  end
-
-  add_index "audits", ["associated_id", "associated_type"], :name => "associated_index"
-  add_index "audits", ["auditable_id", "auditable_type"], :name => "auditable_index"
-  add_index "audits", ["created_at"], :name => "index_audits_on_created_at"
-  add_index "audits", ["user_id", "user_type"], :name => "user_index"
+ActiveRecord::Schema.define(:version => 20120221195345) do
 
   create_table "clients", :force => true do |t|
     t.string   "name"
     t.string   "status"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                         :null => false
+    t.datetime "updated_at",                                         :null => false
     t.string   "guid"
     t.string   "initials"
     t.decimal  "overtime_multiplier", :precision => 10, :scale => 2
@@ -50,8 +29,8 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.integer  "commentable_id"
     t.string   "commentable_type"
     t.integer  "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                     :null => false
+    t.datetime "updated_at",                                     :null => false
     t.boolean  "active"
   end
 
@@ -65,8 +44,8 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.string   "email_address"
     t.string   "phone_number"
     t.integer  "client_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                        :null => false
+    t.datetime "updated_at",                        :null => false
     t.boolean  "receives_email", :default => false
     t.string   "street"
     t.string   "city"
@@ -80,8 +59,8 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.integer  "data_vaultable_id"
     t.string   "data_vaultable_type"
     t.text     "encrypted_data"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",          :null => false
+    t.datetime "updated_at",          :null => false
   end
 
   create_table "file_attachments", :force => true do |t|
@@ -91,22 +70,22 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.string   "attachment_file_file_name"
     t.string   "attachment_file_file_content_type"
     t.integer  "attachment_file_file_size"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                        :null => false
+    t.datetime "updated_at",                        :null => false
     t.boolean  "not_valid"
   end
 
   create_table "git_commits", :force => true do |t|
     t.text     "payload"
     t.integer  "git_push_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  :null => false
+    t.datetime "updated_at",  :null => false
   end
 
   create_table "git_pushes", :force => true do |t|
     t.text     "payload"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
     t.integer  "user_id"
   end
 
@@ -114,15 +93,15 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.string   "github_concernable_type"
     t.integer  "github_concernable_id"
     t.integer  "git_push_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",              :null => false
+    t.datetime "updated_at",              :null => false
   end
 
   create_table "projects", :force => true do |t|
     t.string   "name"
     t.integer  "client_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                                            :null => false
+    t.datetime "updated_at",                                                            :null => false
     t.string   "guid"
     t.decimal  "overtime_multiplier", :precision => 10, :scale => 2
     t.string   "git_repo_name"
@@ -138,15 +117,15 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.string   "name",              :limit => 40
     t.string   "authorizable_type", :limit => 40
     t.integer  "authorizable_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                      :null => false
+    t.datetime "updated_at",                      :null => false
   end
 
   create_table "roles_users", :id => false, :force => true do |t|
     t.integer  "user_id"
     t.integer  "role_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "site_settings", :force => true do |t|
@@ -159,13 +138,30 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.integer  "client_id"
   end
 
+  create_table "taggings", :force => true do |t|
+    t.integer  "tag_id"
+    t.integer  "taggable_id"
+    t.string   "taggable_type"
+    t.integer  "tagger_id"
+    t.string   "tagger_type"
+    t.string   "context",       :limit => 128
+    t.datetime "created_at"
+  end
+
+  add_index "taggings", ["tag_id"], :name => "index_taggings_on_tag_id"
+  add_index "taggings", ["taggable_id", "taggable_type", "context"], :name => "index_taggings_on_taggable_id_and_taggable_type_and_context"
+
+  create_table "tags", :force => true do |t|
+    t.string "name"
+  end
+
   create_table "tickets", :force => true do |t|
     t.integer  "project_id"
     t.string   "priority"
     t.string   "name"
     t.text     "description"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                                        :null => false
+    t.datetime "updated_at",                                                        :null => false
     t.string   "guid"
     t.string   "state"
     t.decimal  "estimated_hours", :precision => 10, :scale => 2
@@ -184,8 +180,8 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                             :null => false
+    t.datetime "updated_at",                                             :null => false
     t.string   "first_name"
     t.string   "last_name"
     t.string   "middle_initial"
@@ -215,8 +211,8 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
   create_table "work_units", :force => true do |t|
     t.text     "description"
     t.integer  "ticket_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                     :null => false
+    t.datetime "updated_at",                                     :null => false
     t.decimal  "hours",           :precision => 10, :scale => 2
     t.boolean  "overtime"
     t.datetime "scheduled_at"
@@ -229,5 +225,7 @@ ActiveRecord::Schema.define(:version => 20120126154403) do
     t.string   "hours_type"
     t.decimal  "effective_hours", :precision => 10, :scale => 2
   end
+
+  add_index "work_units", ["user_id"], :name => "index_work_units_on_user_id"
 
 end